Split test_tasks.py
